adk/pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.5.0",
    "litellm>=1.73.0",
    "mcp>=1.9.4",
]
